Added break time property to availability entity

+++ TutoringSystem/TutoringSystem.Application/Dtos/ReservationDtos/ReservationDetailsDto.cs
@@ -14,7 +14,7 @@ namespace TutoringSystem.Application.Dtos.ReservationDtos
         public long Id { get; set; }
         public double Cost { get; set; }
         public DateTime StartTime { get; set; }
-        public double Duration { get; set; }
+        public int Duration { get; set; }
         public string Description { get; set; }
         public ReservationPlace Place { get; set; }
